14 internal grpc (#24)
* use int32 everywhere

starting pbringing in grpc interfce

* log on errors

* fix proto

* more int32

fininsh grpc interface

* graceful shutdown

grpc running on port

internal grpc server tests

+++ task.go
@@ -14,8 +14,8 @@ type Task struct {
 	Payload   []byte
 	CreatedAt time.Time
 
-	Version          int64
-	DeliveryAttempts int64
+	Version          int32
+	DeliveryAttempts int32
 
 	Priority int32
 }
